+++ test/integration/mouse_pick.cc
@@ -18,6 +18,9 @@
 #include <ignition/math/Vector2.hh>
 #include <ignition/math/Vector3.hh>
 
+#include "gazebo/common/Mesh.hh"
+#include "gazebo/common/MeshManager.hh"
+
 #include "gazebo/gui/Actions.hh"
 #include "gazebo/gui/GLWidget.hh"
 #include "gazebo/gui/GuiIface.hh"
@@ -25,6 +28,8 @@
 
 #include "gazebo/gui/model/ModelEditorPalette.hh"
 #include "gazebo/gui/model/ModelCreator.hh"
+
+#include "gazebo/rendering/RayQuery.hh"
 
 #include "mouse_pick.hh"
 
@@ -143,7 +148,7 @@ void MousePickingTest::Shapes()
   QVERIFY(model02Vis->GetHighlighted());
   QVERIFY(!model03Vis->GetHighlighted());
 
-  // pick the third model - box
+  // pick the third model - cylinder
   pickPt = cam->Project(model03Vis->WorldPose().Pos());
   pt = QPoint(pickPt.X(), pickPt.Y());
   QTest::mouseMove(glWidget, pt);
@@ -409,6 +414,80 @@ void MousePickingTest::Transparency()
   mainWindow->close();
   delete mainWindow;
 }
+
+/////////////////////////////////////////////////
+void MousePickingTest::InvalidMesh()
+{
+  this->resMaxPercentChange = 5.0;
+  this->shareMaxPercentChange = 2.0;
+
+  this->Load("worlds/shapes.world", false, false, false);
+
+  gazebo::gui::MainWindow *mainWindow = new gazebo::gui::MainWindow();
+  QVERIFY(mainWindow != NULL);
+  // Create the main window.
+  mainWindow->Load();
+  mainWindow->Init();
+  mainWindow->show();
+
+  std::string modelName = "box";
+
+  // Get the user camera and scene
+  gazebo::rendering::UserCameraPtr cam = gazebo::gui::get_active_camera();
+  QVERIFY(cam != nullptr);
+  gazebo::rendering::ScenePtr scene = cam->GetScene();
+  QVERIFY(scene != nullptr);
+
+  this->ProcessEventsAndDraw(mainWindow);
+
+  gazebo::rendering::VisualPtr modelVis = scene->GetVisual(modelName);
+  QVERIFY(modelVis != nullptr);
+
+  // get the box mesh
+  gazebo::common::Mesh *mesh = const_cast<gazebo::common::Mesh *>(
+      gazebo::common::MeshManager::Instance()->GetMesh("unit_box"));
+
+  // create invalid submesh and add to unit_box mesh
+  // submesh must have no. of indices that is a multiple of 3 (triangle)
+  gazebo::common::SubMesh *subMesh = new gazebo::common::SubMesh();
+  subMesh->SetName("invalid");
+  subMesh->SetPrimitiveType(gazebo::common::SubMesh::TRIANGLES);
+  subMesh->AddIndex(0);
+  subMesh->AddIndex(1);
+  subMesh->AddIndex(2);
+  subMesh->AddIndex(3);
+  subMesh->AddVertex(ignition::math::Vector3d::Zero);
+  subMesh->AddVertex(ignition::math::Vector3d::Zero);
+  subMesh->AddVertex(ignition::math::Vector3d::Zero);
+  subMesh->AddVertex(ignition::math::Vector3d::Zero);
+  subMesh->AddNormal(ignition::math::Vector3d::UnitX);
+  subMesh->AddNormal(ignition::math::Vector3d::UnitX);
+  subMesh->AddNormal(ignition::math::Vector3d::UnitX);
+  subMesh->AddNormal(ignition::math::Vector3d::UnitX);
+  mesh->AddSubMesh(subMesh);
+
+  // move camera to look at the shapes from +x
+  cam->SetWorldPose(ignition::math::Pose3d(
+      ignition::math::Vector3d(-5, 0.0, 0.5),
+      ignition::math::Quaterniond(0, 0, 0)));
+
+  this->ProcessEventsAndDraw(mainWindow);
+
+  // pick the box
+  auto pickPt = cam->Project(modelVis->WorldPose().Pos());
+
+  gazebo::rendering::RayQuery rayQuery(cam);
+  ignition::math::Vector3d intersect;
+  ignition::math::Triangle3d triangle;
+  QVERIFY(rayQuery.SelectMeshTriangle(pickPt.X(), pickPt.Y(), modelVis,
+      intersect, triangle));
+  QVERIFY(triangle.Valid());
+
+  cam->Fini();
+  mainWindow->close();
+  delete mainWindow;
+}
+
 
 // Generate a main function for the test
 QTEST_MAIN(MousePickingTest)